Add files via upload
Shows connections between the driver U-sub 15 pinholes and the Arduino Uno. For our experiment, we only connected driver A to channel C and B of the first mirror and Driver B to channel A and C of the second mirror. Each channel controls one axis rotation. In our case, for the first mirror, Channel C would reflect an incoming light beam vertically, and Channel B would reflect an incoming light beam horizontally.